Extract output port logic from node component
