+++ app/gui2/src/assets/base.css
@@ -57,6 +57,11 @@
   --right-dock-default-width: 40%;
 }
 
+/* Constants */
+:root {
+  --pi: 3.141592653589793;
+}
+
 *,
 *::before,
 *::after {
