+++ app/gui2/src/components/GraphEditor/GraphNode.vue
@@ -7,6 +7,7 @@ import GraphNodeMessage, {
   iconForMessageType,
   type MessageType,
 } from '@/components/GraphEditor/GraphNodeMessage.vue'
+import GraphNodeOutputPorts from '@/components/GraphEditor/GraphNodeOutputPorts.vue'
 import GraphNodeSelection from '@/components/GraphEditor/GraphNodeSelection.vue'
 import GraphVisualization from '@/components/GraphEditor/GraphVisualization.vue'
 import type { NodeCreationOptions } from '@/components/GraphEditor/nodeCreation'
@@ -15,7 +16,6 @@ import NodeWidgetTree, {
   ICON_WIDTH,
 } from '@/components/GraphEditor/NodeWidgetTree.vue'
 import SvgIcon from '@/components/SvgIcon.vue'
-import { useApproach } from '@/composables/animation'
 import { useDoubleClick } from '@/composables/doubleClick'
 import { usePointer, useResizeObserver } from '@/composables/events'
 import { injectGraphNavigator } from '@/providers/graphNavigator'
@@ -31,19 +31,8 @@ import type { Opt } from '@/util/data/opt'
 import { Rect } from '@/util/data/rect'
 import { Vec2 } from '@/util/data/vec2'
 import { displayedIconOf } from '@/util/getIconName'
-import { setIfUndefined } from 'lib0/map'
 import type { ExternalId, VisualizationIdentifier } from 'shared/yjsModel'
-import type { EffectScope } from 'vue'
-import {
-  computed,
-  effectScope,
-  onScopeDispose,
-  onUnmounted,
-  ref,
-  shallowRef,
-  watch,
-  watchEffect,
-} from 'vue'
+import { computed, onUnmounted, ref, shallowRef, watch, watchEffect } from 'vue'
 
 const MAXIMUM_CLICK_LENGTH_MS = 300
 const MAXIMUM_CLICK_DISTANCE_SQ = 50
@@ -89,12 +78,6 @@ const nodeSelection = injectGraphSelection(true)
 const projectStore = useProjectStore()
 const graph = useGraphStore()
 const navigator = injectGraphNavigator(true)
-
-const outputPortsSet = computed(() => {
-  const bindings = graph.db.nodeOutputPorts.lookup(nodeId.value)
-  if (bindings.size === 0) return new Set([nodeId.value])
-  return bindings
-})
 
 const nodeId = computed(() => asNodeId(props.node.rootExpr.id))
 const potentialSelfArgumentId = computed(() => props.node.primarySubject)
@@ -242,9 +225,11 @@ const isVisualizationFullscreen = computed(() => props.node.vis?.fullscreen ?? f
 
 const bgStyleVariables = computed(() => {
   const { x: width, y: height } = nodeSize.value
+  const visBelowNode = graphSelectionSize.value.y - nodeSize.value.y
   return {
     '--node-width': `${width}px`,
     '--node-height': `${height}px`,
+    '--output-port-transform': `translateY(${visBelowNode}px)`,
   }
 })
 
@@ -373,11 +358,6 @@ function getRelatedSpanOffset(domNode: globalThis.Node, domOffset: number): numb
   return domOffset
 }
 
-const handlePortClick = useDoubleClick(
-  (event: PointerEvent, portId: AstId) => emit('outputPortClick', event, portId),
-  (event: PointerEvent, portId: AstId) => emit('outputPortDoubleClick', event, portId),
-).handleClick
-
 const handleNodeClick = useDoubleClick(
   (e: MouseEvent) => {
     if (!significantMove.value) {
@@ -389,78 +369,6 @@ const handleNodeClick = useDoubleClick(
     if (!significantMove.value) emit('doubleClick')
   },
 ).handleClick
-
-interface PortData {
-  clipRange: [number, number]
-  label: string | undefined
-  portId: AstId
-}
-
-const outputPorts = computed((): PortData[] => {
-  const ports = outputPortsSet.value
-  const numPorts = ports.size
-  return Array.from(ports, (portId, index): PortData => {
-    return {
-      clipRange: [index / numPorts, (index + 1) / numPorts],
-      label: numPorts > 1 ? graph.db.getOutputPortIdentifier(portId) : undefined,
-      portId,
-    }
-  })
-})
-
-const outputHovered = ref<AstId>()
-const anyPortDisconnected = computed(() => {
-  for (const port of outputPortsSet.value) {
-    if (graph.disconnectedEdgePorts.has(port)) return true
-  }
-  return false
-})
-const portsVisible = computed(
-  () =>
-    selectionVisible.value ||
-    (outputHovered.value && outputPortsSet.value.has(outputHovered.value)) ||
-    anyPortDisconnected.value,
-)
-
-const portsHoverAnimation = useApproach(() => (portsVisible.value ? 1 : 0), 50, 0.01)
-
-watchEffect(() => emit('update:hoverAnim', portsHoverAnimation.value))
-
-const hoverAnimations = new Map<AstId, [ReturnType<typeof useApproach>, EffectScope]>()
-watchEffect(() => {
-  const ports = outputPortsSet.value
-  for (const key of hoverAnimations.keys())
-    if (!ports.has(key)) {
-      hoverAnimations.get(key)?.[1].stop()
-      hoverAnimations.delete(key)
-    }
-  for (const port of outputPortsSet.value) {
-    setIfUndefined(hoverAnimations, port, () => {
-      // Because `useApproach` uses `onScopeDispose` and we are calling it dynamically (i.e. not at
-      // the setup top-level), we need to create a detached scope for each invocation.
-      const scope = effectScope(true)
-      const approach = scope.run(() =>
-        useApproach(() => (outputHovered.value === port ? 1 : 0), 50, 0.01),
-      )!
-      return [approach, scope]
-    })
-  }
-})
-
-// Clean up dynamically created detached scopes.
-onScopeDispose(() => hoverAnimations.forEach(([_, scope]) => scope.stop()))
-
-function portGroupStyle(port: PortData) {
-  const [start, end] = port.clipRange
-  const visBelowNode = graphSelectionSize.value.y - nodeSize.value.y
-  return {
-    '--hover-animation': portsHoverAnimation.value,
-    '--direct-hover-animation': hoverAnimations.get(port.portId)?.[0].value ?? 0,
-    '--port-clip-start': start,
-    '--port-clip-end': end,
-    transform: `translateY(${visBelowNode}px`,
-  }
-}
 
 const visRect = shallowRef<Rect>()
 function updateVisualizationRect(rect: Rect | undefined) {
@@ -615,20 +523,13 @@ watchEffect(() => {
     />
     <svg class="bgPaths" :style="bgStyleVariables">
       <rect class="bgFill" />
-      <template v-for="port of outputPorts" :key="port.portId">
-        <g :style="portGroupStyle(port)">
-          <g class="portClip">
-            <rect
-              class="outputPortHoverArea"
-              @pointerenter="outputHovered = port.portId"
-              @pointerleave="outputHovered = undefined"
-              @pointerdown.stop.prevent="handlePortClick($event, port.portId)"
-            />
-            <rect class="outputPort" />
-          </g>
-          <text class="outputPortLabel">{{ port.label }}</text>
-        </g>
-      </template>
+      <GraphNodeOutputPorts
+        :nodeId="nodeId"
+        :forceVisible="selectionVisible"
+        @portClick="(...args) => emit('outputPortClick', ...args)"
+        @portDoubleClick="(...args) => emit('outputPortDoubleClick', ...args)"
+        @update:hoverAnim="emit('update:hoverAnim', $event)"
+      />
     </svg>
   </div>
 </template>
@@ -647,68 +548,6 @@ watchEffect(() => {
   --output-port-hovered-extra-width: 2px;
   --output-port-overlap: -8px;
   --output-port-hover-width: 20px;
-}
-
-.outputPort,
-.outputPortHoverArea {
-  x: calc(0px - var(--output-port-width) / 2);
-  y: calc(0px - var(--output-port-width) / 2);
-  height: calc(var(--node-height) + var(--output-port-width));
-  width: calc(var(--node-width) + var(--output-port-width));
-  rx: calc(var(--node-border-radius) + var(--output-port-width) / 2);
-
-  fill: none;
-  stroke: var(--node-color-port);
-  stroke-width: calc(var(--output-port-width) + var(--output-port-overlap-anim));
-  transition: stroke 0.2s ease;
-  --horizontal-line: calc(var(--node-width) - var(--node-border-radius) * 2);
-  --vertical-line: calc(var(--node-height) - var(--node-border-radius) * 2);
-  --radius-arclength: calc(
-    (var(--node-border-radius) + var(--output-port-width) * 0.5) * 2 * 3.141592653589793
-  );
-
-  stroke-dasharray: calc(var(--horizontal-line) + var(--radius-arclength) * 0.5) 10000%;
-  stroke-dashoffset: calc(
-    0px - var(--horizontal-line) - var(--vertical-line) - var(--radius-arclength) * 0.25
-  );
-  stroke-linecap: round;
-}
-
-.outputPort {
-  --output-port-overlap-anim: calc(var(--hover-animation) * var(--output-port-overlap));
-  --output-port-width: calc(
-    var(--output-port-max-width) * var(--hover-animation) + var(--output-port-hovered-extra-width) *
-      var(--direct-hover-animation) - var(--output-port-overlap-anim)
-  );
-  pointer-events: none;
-}
-
-.outputPortHoverArea {
-  --output-port-width: var(--output-port-hover-width);
-  stroke-width: var(--output-port-hover-width);
-  stroke: transparent;
-  /* Make stroke visible to debug the active area: */
-  /* stroke: red; */
-  stroke-linecap: butt;
-  pointer-events: stroke;
-  cursor: pointer;
-}
-
-.portClip {
-  clip-path: inset(
-    0 calc((1 - var(--port-clip-end)) * (100% + 1px) - 0.5px) 0
-      calc(var(--port-clip-start) * (100% + 1px) + 0.5px)
-  );
-}
-
-.outputPortLabel {
-  user-select: none;
-  pointer-events: none;
-  z-index: 10;
-  text-anchor: middle;
-  opacity: calc(var(--hover-animation) * var(--hover-animation));
-  fill: var(--node-color-primary);
-  transform: translate(50%, calc(var(--node-height) + var(--output-port-max-width) + 16px));
 }
 
 .bgFill {
